migrate getFocusKeybinding
