+++ clients/vscode/src/chat/ChatPanelViewProvider.ts
@@ -299,7 +299,7 @@ export class ChatPanelViewProvider {
     if (serverInfo.config.token) {
       this.client?.cleanError();
 
-      const focusKeybinding = await this.getFocusKeybinding();
+      const focusKeybinding = await this.webviewHelper.getFocusKeybinding();
       getLogger().info("focus key binding: ", focusKeybinding);
 
       this.client?.init({
@@ -309,19 +309,6 @@ export class ChatPanelViewProvider {
         focusKey: focusKeybinding,
       });
     }
-  }
-
-  private async getFocusKeybinding(): Promise<FocusKeybinding | undefined> {
-    const focusCommand = "tabby.chatView.focus";
-    const defaultFocusKey = contributes.keybindings.find((cmd) => cmd.command === focusCommand);
-    const defaultKeybinding = defaultFocusKey
-      ? parseKeybinding(this.isMac && defaultFocusKey.mac ? defaultFocusKey.mac : defaultFocusKey.key)
-      : undefined;
-
-    const allKeybindings = await readUserKeybindingsConfig();
-    const userShortcut = allKeybindings?.find((keybinding) => keybinding.command === focusCommand);
-
-    return userShortcut ? parseKeybinding(userShortcut.key) : defaultKeybinding;
   }
 
   private displayDisconnectedPage() {
